Cypress' recommended best practice for items created during tests is to clear them before the tests run and then leave them after the tests, so that devs can easily investigate items that failed the test.

- **index.ts** - new before() hook that will check the ENV of the baseUrl and use it to set a Cypress.env "deleteItemsAfter", if that variable is not already included in the env object of cypress.json. This env variable will inform all tests that create items if the items should be deleted afterwards or not. Currently only PRF will set deleteItemsAfter to true.
- **setupCommands.ts** - new commands that will query an endpoint for all Cypress items and delete them. These commands will not fail the tests if the query fails or a deletion is not successful, since this command is used in a before() hook.
- **All update mutation tests** - Changed so that a new items is created before each test. Previously each file would create only one item and use it for all tests, which had too much potential for tests interfering with each other.
- Some changes to tests that interact with the storefront, to improve the process of Cypress switching from one baseUrl to the default one, which keeps Cypress from crashing.

Related work items: #211254

+++ src/ecomm-api-svc/cypress/integration/mutations/category/updateCategory.spec.ts
@@ -5,9 +5,11 @@ import { confirmStorefrontEnvValues, createInfoDummy, SupplementalItemRecord, to
 // TEST COUNT: 16
 describe('Mutation: updateCategory', () => {
     var id = '';
-    var updateCount = 0;
+    var updateCount = 0;    // TODO: Appraise whether this is really useful or not
+    var itemCount = 1;
     var extraIds = [] as SupplementalItemRecord[];
     const mutationName = 'updateCategory';
+    const deleteMutName = "deleteCategory";
     const queryName = "categories";
     const itemPath = 'category';
     const infoName = "categoryInfo";
@@ -39,25 +41,37 @@ describe('Mutation: updateCategory', () => {
         });
     };
 
+    var deleteItemsAfter = undefined as boolean | undefined;
     before(() => {
+        deleteItemsAfter = Cypress.env("deleteItemsAfter");
+        cy.deleteCypressItems(queryName, deleteMutName, infoName);
+    });
+
+    beforeEach(() => {
         // Create an item for the tests to update
-        const name = `Cypress ${mutationName} Test`;
+        const name = `Cypress ${mutationName} Test #${itemCount}`;
         const input = `{${infoName}: [{name: "${name}", description: "Cypress testing for ${mutationName}", languageCode: "Standard"}] }`;
         cy.createAndGetId(createName, itemPath, input).then((returnedId: string) => {
             assert.exists(returnedId);
             id = returnedId;
+            itemCount++;
             parentCatName = name; // Only going to use this in one test, but we still need it.
         });
-    });
+    })
 
-    after(() => {
+    afterEach(() => {
+        if (!deleteItemsAfter) {
+            return;
+        }
         if (originalBaseUrl !== "" && Cypress.config("baseUrl") !== originalBaseUrl) {
             Cypress.log({message: "Switching the baseUrl back to the original"});
             Cypress.config("baseUrl", originalBaseUrl);
             cy.wait(1000);
         }
         // Delete any supplemental items we created
-        cy.deleteSupplementalItems(extraIds);
+        cy.deleteSupplementalItems(extraIds).then(() => {
+            extraIds = [];
+        });
 
         // Delete the child and parent category
         cy.deleteParentAndChildCat({name: childCatName, id: childCatId}, parentCatName, parentCatId);
@@ -67,7 +81,7 @@ describe('Mutation: updateCategory', () => {
 
         if (id !== "") {
             // Delete the item we've been updating
-            cy.deleteItem("deleteCategory", id);
+            cy.deleteItem(deleteMutName, id);
         }
     });
 
@@ -225,7 +239,7 @@ describe('Mutation: updateCategory', () => {
             cy.createAndGetId(createName, itemPath, input, "customData").then((createdItem) => {
                 assert.exists(createdItem.id);
                 assert.exists(createdItem.customData);
-                extraIds.push({itemId: createdItem.id, deleteName: "deleteCategory", itemName: info[0].name, queryName: queryName});
+                extraIds.push({itemId: createdItem.id, deleteName: deleteMutName, itemName: info[0].name, queryName: queryName});
                 const newInfo = [{name: `Cypress ${mutationName} CD extra updated`, description: `${mutationName} CD cypress test`, languageCode: "Standard"}];
                 const newCustomData = {data: `${itemPath} customData`, newDataField: { canDelete: true }};
                 const mutation = `mutation {
@@ -565,6 +579,7 @@ describe('Mutation: updateCategory', () => {
         });
     });
    
+    // TODO: Move this to a category file in the misc folder
     context.skip("Testing in storefront", () => {
         // This cannot be run on its own without another test run before it.
         // The baseUrl changes too fast for us to save it as originalBaseUrl if it's run on its own. This prevents us from making API calls
@@ -577,7 +592,7 @@ describe('Mutation: updateCategory', () => {
             const pageSize = 10;
             const input = {categoryInfo: [{name: name, languageCode: "Standard"}], published: published, displayOrder: displayOrder, pageSize: pageSize}
             cy.createAndGetId(createName, itemPath, toFormattedString(input), undefined, originalBaseUrl).then((returnedId: string) => {
-                extraIds.push({itemId: returnedId, deleteName: "deleteCategory", itemName: name, queryName: queryName});
+                extraIds.push({itemId: returnedId, deleteName: deleteMutName, itemName: name, queryName: queryName});
                 name = `Cypress TopMenu Update ${Cypress._.random(0, 999)}`;
                 const newInfo = [{name: name, languageCode: "Standard"}];
                 const showInTopMenu = true;
