+++ src/ecomm-api-svc/cypress/integration/queries/categories.spec.ts
@@ -152,8 +152,20 @@ describe('Query: categories', () => {
                 }
             });
         };
+        
+        var deleteItemsAfter = undefined as boolean | undefined;
+        before(() => {
+            deleteItemsAfter = Cypress.env("deleteItemsAfter");
+            cy.deleteCypressItems(queryName, deleteName, infoName).then(() => {
+                cy.deleteCypressItems(productQuery, productDelete, "productInfo", `Cypress ${queryName}`);
+            });
+        });
+
         // Ensure deletion of the items we created for the productId test
         after(() => {
+            if (!deleteItemsAfter) {
+                return;
+            }
             cy.deleteSupplementalItems(createdProducts);
             cy.deleteSupplementalItems(createdItems);
         });
