+++ src/ecomm-api-svc/cypress/integration/queries/manufacturers.spec.ts
@@ -154,8 +154,19 @@ describe('Query: manufacturers', () => {
             });
         };
 
+        var deleteItemsAfter = undefined as boolean | undefined;
+        before(() => {
+            deleteItemsAfter = Cypress.env("deleteItemsAfter");
+            cy.deleteCypressItems(queryName, deleteName, infoName).then(() => {
+                cy.deleteCypressItems(productQuery, productDelete, "productInfo", `Cypress ${queryName}`);
+            });
+        });
+
         // Ensure deletion of the items we created for the productId test
         after(() => {
+            if (!deleteItemsAfter) {
+                return;
+            }
             cy.deleteSupplementalItems(createdProducts);
             cy.deleteSupplementalItems(createdItems);
         });
